milestone 28 - the last mile - 4 of 4
* array refactor: Array(i64) vs. Array(n)
* finalized Expression::to_pure()
* added support for binary, octal and hexadecimal literals
* added util::to_binary(a)
* added FieldMetadata::is_external
* added BLOBStore to enable support for variable-length cells
* added HybridRowCollection (binary and disk)
* added Tuples
* removed FileEmbeddedRowCollection
* websockets handler refactor

+++ src/core/numbers.rs
@@ -713,7 +713,7 @@ impl Neg for Numbers {
         match self {
             Ack => I64Value(0),
             RowId(n) => I64Value(-(n as i64)),
-            RowsAffected(n) => I64Value(-(n as i64)),
+            RowsAffected(n) => I64Value(-n),
             DateValue(d) => I64Value(-d),
             F32Value(n) => F32Value(-n),
             F64Value(n) => F64Value(-n),
